requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
airportsdata==20241001
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens
async-timeout==5.0.1
attrs==24.2.0
banal==1.0.6
bert-score==0.3.13
bitsandbytes==0.44.1
blis==1.0.1
breadability==0.1.20
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
cmake==3.31.1
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
dataset==1.6.2
datasets==2.16.0
debugpy
decorator
dill==0.3.6
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
entrypoints
evaluate==0.4.3
exceptiongroup 
executing
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
FlagEmbedding==1.2.11
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2023.10.0
gradio==5.8.0
gradio_client==1.5.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
interegular==0.3.3
ipdb==0.13.13
jedi
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpickle==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
lm-format-enforcer==0.10.1
longdocfactscore==1.0.0
lxml==5.3.0
Mako==1.3.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matching==1.4
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.14
murmurhash==1.0.11
nest_asyncio
networkx==3.4.2
ninja==1.11.1.2
nltk==3.6.2
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
openai==0.28.1
orjson==3.10.12
outlines==0.1.7
outlines_core==0.1.17
packaging
pandas==2.2.3
parso
peft==0.13.2
pexpect
pickleshare
pillow==11.0.0
platformdirs
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
prompt_toolkit
propcache==0.2.1
protobuf==5.29.0
psutil
ptyprocess
pure_eval
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyarrow-hotfix==0.6
pycountry==24.6.1
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments
pyparsing==3.2.0
pysbd==0.3.4
python-dateutil
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.1
ruff==0.8.3
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.4.8
stack_data
starlette==0.41.3
sumy==0.11.0
sympy==1.13.3
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.2
torch==2.3.0
tornado
tqdm==4.67.1
traitlets
transformers==4.46.3
triton==2.3.0
typer==0.15.0
typing_extensions
tzdata==2024.2
Unidecode==1.2.0
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.0
wcwidth
weasel==0.4.1
websockets==14.1
wrapt==1.17.0
xformers==0.0.26.post1
xxhash==3.5.0
yarl==1.18.3
